updated for version 7.4.023 Problem:    Compiler warning on 64 bit windows. Solution:   Add type cast. (Mike Williams)

+++ src/edit.c
@@ -5189,7 +5189,7 @@ ins_complete(c)
 	    mb_ptr_back(line, p);
 	    while (vim_isfilec(PTR2CHAR(p)) && p >= line)
 		mb_ptr_back(line, p);
-	    startcol = p - line;
+	    startcol = (int)(p - line);
 
 	    compl_col += ++startcol;
 	    compl_length = (int)curs_col - startcol;
